+++ src/main/java/com/netcracker/dao/impl/mapper/AccountIncomeMapper.java
@@ -17,9 +17,9 @@ public class AccountIncomeMapper implements RowMapper<AccountIncome> {
         AbstractAccountOperation accountIncome =
                 new AccountIncome.Builder()
                 .accountId(resultSet.getBigDecimal("account_income_id").toBigInteger())
-                .accountAmount(resultSet.getBigDecimal("income_amount").toBigInteger())
+                .accountAmount(resultSet.getLong("income_amount"))
                 .accountDate(resultSet.getDate("date").toLocalDate())
-                .accountUserId(resultSet.getLong("user_id"))
+                .accountUserId(resultSet.getBigDecimal("user_id").toBigInteger())
                 .build();
         return (AccountIncome) accountIncome;
     }
